consensus: bug 23171: removing references to dotnet v3.5
A couple of our .csproj file contained references to v3.5 of dotnet,
but we expect to use v4.5 for everything we build. 

This setting in PacBio.Analysis.ConsensusCoreManaged.csproj is the
biggest troublemaker:	

  <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>

That actually forces xbuild to fire off a different compiler (gmcs) 
instead of mcs (underneath, gmcs is just "mcs -sdk:2", forcing it 
to use dotnet 2.0 assemblies).  

But our mono_fsharp prebuilt.out install has all the 2.0 assemblies
trimmed out, so it probably builds with 4.0/4.5 anyway.  But we should
make sure all our .csproj files are consistent, so that they at least
fire up the same compiler.

Also removed some other suspect settings (which didn't seem to make 
a difference in the compile), such as:

  <Reference Include="System.Core">
      <RequiredTargetFramework>3.5</RequiredTargetFramework>
  </Reference>

  <OldToolsVersion>3.5</OldToolsVersion>

[git-p4: depot-paths = "//depot/software/common/ConsensusCore/": change = 124678]