consensus: bug 23171: add "if linux" to CSharp.mk
Make sure to only run the os detection shell command if we are
running under linux to avoid error on other oses that may be using
CSharp.mk.  This still assumes that the csharp-pacbio target is only
used by the smrtanalysis build (not completely sure that is a valid
assumption though).

[git-p4: depot-paths = "//depot/software/common/ConsensusCore/": change = 124692]