consensus: bug 23171: fixing the primary build
The assumption that the primary build did not call the csharp-pacbio
target was incorrect, so the primary build was breaking trying to
point to the new multi-os version of swig for the smrtanalysis build.

We now detect if we are running a smrtanalysis build (by the existance
of the software/smrtanalysis directory) and set the location of
the swig binary and libs accordingly.

[git-p4: depot-paths = "//depot/software/common/ConsensusCore/": change = 125017]