QuiverConfigTable: Add a "fallback" behavior
If the chemistry isn't found in the list, we will look for "*" which
is the fallback model.  In practice I will only be using this in cases
where the user overrides the model setting---but I still want
ConsensusCore to see the real chemistry names.

This isn't the cleanest solution from a SW point of view but it is
simple and has some virtues in allowing client code to be decoupled.

[Tracking bug 23893]

[git-p4: depot-paths = "//depot/software/common/ConsensusCore/": change = 131250]