Fix MappedConsensus slowness.  Quiver and MC are now about the same speed on a cmp.h5.
The slowness was due to the huge bands used by LAA and
CCS (ScoreDiff=18) vs Quiver (ScoreDiff=6).  I would really like all
of our code to be using the same banding but that is risky for 2.3.
We will revisit post 2.3.

This required adjusting ConsensusCore API and I brought method names
in the class up to Pascal-case standard (style guide) while in there,
so I will have to fix GenomicConsensus in a subsequent checkin.

[Tracking bug 24842]

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/cpp/ConsensusCore/": change = 139279]